requirements.txt
numpy>=1.24,<2
Pillow>=10.2,<10.3
scikit-learn>=1.3
joblib>=1.3
facenet-pytorch>=2.5
torch>=2.1
opencv-python>=4.8
mediapipe==0.10.9
